Use fixed position for the active card tagline
Now that the div.card can scroll, this allows the
taglines to scroll with the card content because they
were originally absolutely positioned with respected to
the fixed div.card

Instead, make them fixed to the viewport.

Since the card content can scroll too, we have to actually set the
card heights when we allow the card to scroll so that the .tagline
will be properly placed.

This should probably be set at the hoc.css level rather than the
demo.css level, but leaving it there for now to be minimally invasive.