Adjust tagline positions
Need to adjust these since the taglines are popping in and out of the
dom.

this makes it a big fragile and forces the 2 css files and the hoc.js
to be tightly coupled by the padding/top/bottom/left/right values.

ideally pull these out and refactor the hoc.css/demo.css to pull all the
framework specific stuff into hoc.css and the individual styles into
demo.css.

hoc.js also has a hard set 35px offset that is related to the
bottom:35px css in order to make the tagline scrolling look smooth :/